+++ src/regional_settings.h
@@ -288,6 +288,7 @@ struct overmap_highway_settings {
 
     int longest_slant_length = 0;
     int longest_bend_length = 0;
+    int HIGHWAY_MAX_DEVIANCE = 0;
 
     bool needs_finalize = false;
     void finalize();
